# src/main/java/nl/demo/mail/mailservice/model/Attach.java
package nl.demo.mail.mailservice.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

import java.io.File;
import java.util.List;

@Getter
@Setter
@ToString
@NoArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Attach {

    // TODO: File -> Binary!!!
    private List<File> file;
//    private List<Binary> file;

}