# src/test/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger UI van FCD Emailen (generated by Openapi Generator https://github.com/openapitools/openapi-generator)",
    "version": "1.0.0",
    "title": "FCD Emailen",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/v1",
  "tags": [
    {
      "name": "fcd-emailen",
      "description": "API voor het versturen van mail"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/mail": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Maak mail aan:",
        "operationId": "addMail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Mail object dat moet aangemaakt worden",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Mail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "mail"
        ],
        "summary": "Wijzig mail aan:",
        "operationId": "updateMail",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Mail object dat moet aangemaakt worden",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Mail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Mail Not Found"
          },
          "405": {
            "description": "Validation Exception"
          }
        }
      }
    },
    "/mail/{mailId}": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "Find mail by ID",
        "description": "Returns a single mail",
        "operationId": "getMailById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mailId",
            "in": "path",
            "description": "ID of mail to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Mail"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Mail not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Mail": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "header": {
          "$ref": "#/definitions/Header"
        },
        "body": {
          "$ref": "#/definitions/Body"
        },
        "attach": {
          "$ref": "#/definitions/Attach"
        }
      }
    },
    "Header": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string",
          "format": "mail"
        },
        "replyto": {
          "type": "string",
          "format": "mail"
        },
        "to": {
          "type": "string",
          "format": "mail"
        },
        "cc": {
          "type": "string",
          "format": "mail"
        },
        "bcc": {
          "type": "string",
          "format": "mail"
        },
        "subject": {
          "type": "string"
        }
      }
    },
    "Body": {
      "type": "string"
    },
    "Attach": {
      "type": "object"
    }
  }
}